+++ tests/test_measurements.py
@@ -592,7 +592,6 @@ class TestWiresInExpval:
         assert np.allclose(circuit1(), circuit2(), atol=tol)
 
 
-@pytest.mark.skipif(ld._new_API, reason="Old API required")
 class TestSample:
     """Tests that samples are properly calculated."""
 
@@ -608,28 +607,35 @@ class TestSample:
         """Tests if the samples returned by the sample function have
         the correct dimensions
         """
-        dev = qubit_device(wires=2)
-
-        dev.apply([qml.RX(1.5708, wires=[0]), qml.RX(1.5708, wires=[1])])
-
-        dev.shots = shots
-        dev._wires_measured = wires
-        dev._samples = dev.generate_samples()
-        s1 = dev.sample(qml.PauliZ(wires=[0]))
-        assert np.array_equal(s1.shape, (dev.shots,))
+        dev = qubit_device(wires=2, shots=shots)
+        ops = [qml.RX(1.5708, wires=[0]), qml.RX(1.5708, wires=[1])]
+        obs = qml.PauliZ(wires=[0])
+        if ld._new_API:
+            tape = qml.tape.QuantumScript(ops, [qml.sample(op=obs)], shots=shots)
+            s1 = dev.execute(tape)
+        else:
+            dev.apply(ops)
+            dev._wires_measured = wires
+            dev._samples = dev.generate_samples()
+            s1 = dev.sample(obs)
+        assert np.array_equal(s1.shape, (shots,))
 
     def test_sample_values(self, qubit_device, tol):
         """Tests if the samples returned by sample have
         the correct values
         """
-        dev = qubit_device(wires=2)
-
-        dev.shots = 1000
-        dev.apply([qml.RX(1.5708, wires=[0])])
-        dev._wires_measured = {0}
-        dev._samples = dev.generate_samples()
-
-        s1 = dev.sample(qml.PauliZ(0))
+        shots = 1000
+        dev = qubit_device(wires=2, shots=shots)
+        ops = [qml.RX(1.5708, wires=[0])]
+        obs = qml.PauliZ(0)
+        if ld._new_API:
+            tape = qml.tape.QuantumScript(ops, [qml.sample(op=obs)], shots=shots)
+            s1 = dev.execute(tape)
+        else:
+            dev.apply(ops)
+            dev._wires_measured = {0}
+            dev._samples = dev.generate_samples()
+            s1 = dev.sample(qml.PauliZ(0))
 
         # s1 should only contain 1 and -1, which is guaranteed if
         # they square to 1
@@ -683,7 +689,6 @@ class TestWiresInVar:
 
 
 @flaky(max_runs=5)
-@pytest.mark.skipif(ld._new_API, reason="Old API required")
 @pytest.mark.parametrize("shots", [10000, [10000, 11111]])
 @pytest.mark.parametrize("measure_f", [qml.counts, qml.expval, qml.probs, qml.sample, qml.var])
 @pytest.mark.parametrize(